+++ src/org/bds/test/unit/TestCasesFile.java
@@ -1,0 +1,17 @@
+package org.bds.test.unit;
+
+import org.bds.test.TestCasesBase;
+
+/**
+ * Test cases for file methods
+ *
+ * @author pcingola
+ */
+public class TestCasesFile extends TestCasesBase {
+
+    public TestCasesFile() {
+        dir = "test/unit/run/";
+    }
+
+
+}
