+++ src/org/bds/test/unit/TestCasesFunctionsMethodsNative.java
@@ -1,0 +1,17 @@
+package org.bds.test.unit;
+
+import org.bds.test.TestCasesBase;
+
+/**
+ * Test cases for native functions and methods
+ *
+ * @author pcingola
+ */
+public class TestCasesFunctionsMethodsNative extends TestCasesBase {
+
+    public TestCasesFunctionsMethodsNative() {
+        dir = "test/unit/run/";
+    }
+
+
+}
