+++ src/org/bds/test/unit/TestCasesTryCatchFinally.java
@@ -1,0 +1,17 @@
+package org.bds.test.unit;
+
+import org.bds.test.TestCasesBase;
+
+/**
+ * Test cases for 'try/catch/finally' statements
+ * 
+ * @author pcingola
+ */
+public class TestCasesTryCatchFinally extends TestCasesBase {
+
+    public TestCasesTryCatchFinally() {
+        dir = "test/unit/run/";
+    }
+
+
+}
